backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.8.0
pydantic-settings>=2.3.1
sqlalchemy>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0
pgvector>=0.2.5
langchain>=0.2.1
langchain-aws>=0.1.6
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
